requirements.txt
# Core Python dependencies
setuptools>=65.0.0
wheel>=0.37.0
pip>=22.0.0

# Flask and web framework
flask==2.3.3
flask-cors==4.0.0
flask-mail==0.9.1
flask-session==0.5.0

# Database and storage
chromadb==0.4.18
pymongo==4.5.0
dnspython==2.4.2

# Authentication and security
bcrypt==4.0.1
PyJWT==2.8.0

# Utilities
requests==2.31.0
gunicorn==21.2.0
numpy<2.0.0
python-dotenv==1.0.0

# Additional dependencies for production
psycopg2-binary==2.9.7

# Email support
itsdangerous==2.1.2
werkzeug==2.3.7

# AI dependencies
sentence-transformers==2.2.2

# Optional AI dependencies (uncomment if needed)
# openai==1.3.0